Inizio Services

+++ Lumiere-ProjectWork/src/main/java/lumiere/central/repository/AttoreRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import lumiere.central.model.Attore;
 
-public interface AttoreRepository extends JpaRepository <Attore, Integer> {
+public interface AttoreRepository extends JpaRepository <Attore, Long> {
 
 }
