+++ Lumiere-ProjectWork/src/main/java/lumiere/central/repository/GenereRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import lumiere.central.model.Genere;
 
-public interface GenereRepository extends JpaRepository <Genere, Integer> {
+public interface GenereRepository extends JpaRepository <Genere, Long> {
 
 }
